requirements.txt
asgiref==3.5.2
Django==4.1.3
django-debug-toolbar==3.7.0
django-filter==22.1
django-routers==0.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
mysqlclient==2.1.1
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
